+++ mcramp/LEGACY/scat/guide.py
@@ -10,7 +10,7 @@ import re
 class SGuide(SPrim):
     def __init__(self, g_pos=(0., 0., 0.), w1=0, h1=0, w2=0, h2=0, l=0, 
                  R0=0, Qc=0, alpha=0, m=1, W=0, idx=0, ctx=0, max_bounces=5):
-        self.g_pos  = np.array((g_pos[0], g_pos[1], g_pos[2], 0. ), dtype=clarr.vec.float3)
+        self.g_pos  = np.array((g_pos[0], g_pos[1], g_pos[2], 0. ), dtype=clarr.vec.double3)
         self.w1     = np.float64(w1)
         self.h1     = np.float64(h1)
         self.w2     = np.float64(w2)
